Move setup scripts from /config to /app/setup
